requirements.txt
python-gitlab~=5.3.0
openai~=1.59.3